 requirements.txt
buildozer==1.4.0
certifi==2022.9.24
charset-normalizer==2.1.1
Cython==0.29.32
decorator==5.1.1
distlib==0.3.6
docutils==0.19
filelock==3.8.0
idna==3.4
Kivy==2.1.0
kivy-deps.angle==0.3.2
kivy-deps.glew==0.3.1
pypiwin32==223
pywin32==304
requests==2.28.1
self==2020.12.3
sh==1.14.2
urllib3==1.26.12
virtualenv==20.16.5